Derek's original code.
