+++ Battleship/GameState.cs
@@ -1,0 +1,9 @@
+﻿public enum GameState
+{
+    MainMenu,
+    Playing,
+    ShipSelection,
+    Settings,
+    Postgame,
+    Exit
+}
