--- GameState.cs
@@ -1,9 +1,0 @@
-﻿public enum GameState
-{
-    MainMenu,
-    Playing,
-    ShipSelection,
-    Settings,
-    Postgame,
-    Exit
-}
