--- defaultFont.mgcontent
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PipelineBuildEvent xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
-  <SourceFile>C:/Users/derek/source/repos/Project1-Battleship/Battleship/Content/defaultFont.spritefont</SourceFile>
-  <SourceTime>2022-07-26T06:44:10-05:00</SourceTime>
-  <DestFile>C:/Users/derek/source/repos/Project1-Battleship/Battleship/Content/bin/DesktopGL/defaultFont.xnb</DestFile>
-  <DestTime>2024-09-15T04:01:47.4473653-05:00</DestTime>
-  <Importer>FontDescriptionImporter</Importer>
-  <ImporterTime>2022-07-26T06:44:56-05:00</ImporterTime>
-  <Processor>FontDescriptionProcessor</Processor>
-  <ProcessorTime>2022-07-26T06:44:56-05:00</ProcessorTime>
-  <Parameters>
-    <Key>PremultiplyAlpha</Key>
-    <Value>True</Value>
-  </Parameters>
-  <Parameters>
-    <Key>TextureFormat</Key>
-    <Value>Compressed</Value>
-  </Parameters>
-  <Dependencies />
-  <BuildAsset />
-  <BuildOutput />
-</PipelineBuildEvent>
